Create cli.py
